cambios db

+++ proyecto-clinica/src/main/java/SegundaEntrega/persistencia/AsociadoDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface AsociadoDAO {
     AsociadoDTO getAsociadobyDNI(String dni);
     List<AsociadoDTO> getAllAsociados();
-    void darDeBajaAsociado(long id);
+    void darDeBajaAsociado(int id);
     AsociadoDTO createAsociado(AsociadoDTO asociado);
     
 }
